% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxplotBooleanSubsetcellProportions.R
\name{boxplot.boolean.subset.proportions}
\alias{boxplot.boolean.subset.proportions}
\title{Boxplot of the difference in cell proportions for a given condition}
\usage{
boxplot(path, outdir = NULL,
       conditioncol, exp, ctrl, conditioncol2 = ".", parentsubset, boolsubset,
       ylimits = NULL, shortenTitle = FALSE, uniqueSamplesCol)
}
\arguments{
\item{outdir}{(Optional), saves image in output directory if given}

\item{ylimits}{(Optional), yaxis limits. numeric vector}

\item{shortenTitle}{(Optional), a Boolean which specifies whether to shorten the title}

\item{path:}{path to directory holding GatingSetList}

\item{conditioncol:}{name of the column that defines the main experimental condition, e.g. Antigen}

\item{exp:}{experimental value in conditioncol, e.g. ESAT-6}

\item{ctrl:}{control value in conditioncol, e.g. DMSO}

\item{conditioncol2:}{condition on which to stratify box plots}

\item{parentsubset:}{unique name of parent node to use for plots}

\item{boolsubset:}{the full boolean subset to be used by booleanfilter()}

\item{xaxis:}{a marker name to plot on the x-axis}

\item{yaxis:}{a marker name to plot on the y-axis}

\item{uniqueSamplesCol:}{pData column which defines unique samples, e.g. "Time.PTID"}
}
\value{
Boxplot of the difference in cell proportions, unless outdir is given
}
\description{
This function creates a boxplot of the difference in cell proportions
for a given condition (e.g. Antigen) across another variable (e.g. Time)
for the specified boolean subset
}
\examples{
\dontrun{
boxplot.boolean.subset.proportions(path="/home/GatingSetListAllBatches",
                                   conditioncol="Antigen",
                                   exp="ESAT-6",
                                   ctrl="DMSO",
                                   conditioncol2="Time",
                                   parentsubset="8+",
                                   boolsubset="8+/TNFa+&!8+/IFNg+&!8+/IL2+&!8+/IL4+",
                                   ylimits=c(0, 0.013),
                                   uniqueSamplesCol="Time.PTID")
                                   }
}
\keyword{boolean}
\keyword{boxplot}
\keyword{count}
\keyword{subset}
