% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCountsQC.R
\name{boxplot.cell.counts}
\alias{boxplot.cell.counts}
\title{Plot cell counts for each sample (default CD3+ cells)}
\usage{
boxplot.cell.counts <- function(flowJoXmlPath=NULL, gatingSetPath=NULL,
                    fcsPath=if (!.isnull(flowJoXmlPath)) { dirname(flowJoXmlPath) },
                    outdir=NULL, sampleGroup=3, subpopulation="3+",
                    keywords2import=NULL, stratifyByLevel1, stratifyByLevel2=NULL)
}
\arguments{
\item{flowJoXmlPath}{full path to FlowJo Xml file for import (flowJoXmlPath or gatingSetPath or gatingSet is required)}

\item{gatingSetPath}{path to saved GatingSet directory for import (flowJoXmlPath or gatingSetPath or gatingSet is required)}

\item{gatingSet}{a GatingSet object (flowJoXmlPath or gatingSetPath or gatingSet is required)]}

\item{fcsPath}{(Optional) directory where FCS files reside, if not in the same directory as flowJoXmlPath}

\item{outdir}{(Optional) directory to save boxplot in. If not given, returns the boxplot instead.}

\item{sampleGroup}{(Optional) Specify this if you are providing a flowJoXmlPath and the flowJo sample group is not 3}

\item{subpopulation}{(Optional) node name to use for counts, if not 3+}

\item{keywords2import}{(Optional) keywords to import from flowJo into GatingSet metadata, required for flowJoXmlPath}

\item{stratifyByLevel1}{Required keyword on which to stratify boxplot data (usually "PATIENT ID")}

\item{stratifyByLevel2}{(Optional) additional keyword on which to stratify boxplot data}

\item{batch}{the batch name, required if data is passed in as a gatingSet}

\item{threshold}{(Optional) where to draw the threshold cutoff line for number of cells, default 25,000}

\item{yUpperExpand}{(Optional) This will expand the yaxis to have this upper limit, if applicable.}
}
\value{
Boxplot of cell counts stratified by stratifyByLevel1, unless outdir is specified.
}
\description{
Plot the number of cells that are CD3+ (or your specified population) for each sample.
Run this function once for each batch to be combined later. Can use to look for batch effects.
}
\examples{
\dontrun{
boxplot.cell.counts(flowJoXmlPath="/home/Batch1Directory/Batch1FlowJo.xml",
                    keywords2import=c("PATIENT ID", "Barcode"),
                    stratifyByLevel1="PATIENT ID",
                    stratifyByLevel2="Barcode")
                    }
}
\keyword{QC}
\keyword{counts}
