% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CompassStratifiedBackgroundCorrectedProportionsBoxplots.R
\name{boxplotsCompassSubsetBgCorrPropStratified}
\alias{boxplotsCompassSubsetBgCorrPropStratified}
\title{Create Stratified Boxplots of the Background Corrected Proportions for COMPASS subsets}
\usage{
boxplotsCompassSubsetBgCorrPropStratified(compass.subset.comparisons.result,
  cr_cats, stratifyBy, sampleIDCol, parentSubset, themeBaseSize = 18,
  removeGridAndBg = FALSE, showSignificanceBracket = TRUE,
  pvalue_fontsize = NULL, axestitle_fontsize = NULL,
  axestick_fontsize = NULL, font = NULL, geom_jitter_width = 0.05,
  xaxis_title = stratifyBy, dot_size = 0.75, legend_position = c(1, 1),
  show_points = F, padj_method = "bonferroni", ylim = NULL,
  p_alpha = 0.05, stratifyBy_colors = NULL, box_fill_alpha = 0.7,
  legend_fontsize = NULL)
}
\arguments{
\item{compass.subset.comparisons.result}{The output from a run of COMPASSHelpers::compass.subset.comparisons(). Contains the p-values and background-corrected proportions for each subset.}

\item{cr_cats}{a COMPASSResult categories matrix (cr$fit$categories) with rows defining the order in which you want the boxplots to appear. Can be obtained by a call to COMPASSHelpers::orderHeatmapColumnsByCytokinePresence(), for example.}

\item{stratifyBy}{what metadata column to stratify the boxplots by, e.g. "Status"}

\item{sampleIDCol}{e.g. "PATIENT ID"}

\item{parentSubset}{e.g. "4+"}

\item{ylim}{e.g. c(0,0.01)}

\item{p_alpha}{The alpha level to use for choosing which subsets to plot p-values for}

\item{stratifyBy_colors}{e.g. c("RSTR" = "white", "LTBI" = "black")}
}
\description{
Subsets will be plotted in the order in which they appear in the cr_cats matrix.
Only subsets which appear in compass.subset.comparisons.result will be plotted.
Intended to be shown with the categories legend from the corresponding COMPASS heatmap displayed below.
}
\details{
This function is hacked together.
}
\examples{
\dontrun{
compass.subset.comparisons.result <- compass.subset.comparisons(compassResultOrPath=CD4PP1CompassResult,
                                                             gsOrGsListOrPath=gs,
                                                             parentSubset="4+",
                                                             antigenCol="Antigen",
                                                             stimAntigen="Peptide Pool 1",
                                                             controlAntigen="DMSO",
                                                             stratifyBy="Status",
                                                             stratifyByValueMinuend = "LTBI",
                                                             stratifyByValueSubtrahend = "RSTR",
                                                             sampleIDCol="PATIENT ID")
cr_cats <- orderHeatmapColumnsByCytokinePresence(CD4PP1CompassResult, "IFNg",
                                      cats_df_cytokine_order_override = c("CD154", "IL2", "TNF", "CD107a", "IL4", "IL17a", "IFNg"))$fit$categories
boxplotsCompassSubsetBgCorrPropStratified(compass.subset.comparisons.result,
                                          cr_cats,
                                          stratifyBy="Status",
                                          sampleIDCol,
                                          parentSubset,
                                          removeGridAndBg=T,
                                          ylim=c(0, 0.01),
                                          stratifyBy_colors=c("RSTR" = "white", "LTBI" = "black"))
}
}
