% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompassSubsetComparisons.R
\name{compass.subset.comparisons}
\alias{compass.subset.comparisons}
\title{COMPASS Subset Comparisons}
\usage{
compass.subset.comparisons(compassResultOrPath, gsOrGsListOrPath,
  cytokineOfInterestGate, parentSubset, threshold = 0.01, antigenCol,
  stimAntigen, controlAntigen, stratifyBy, outdir = NULL,
  stratifyByValueMinuend = NULL, stratifyByValueSubtrahend = NULL,
  sampleIDCol)
}
\arguments{
\item{compassResultOrPath}{}

\item{gsOrGsListOrPath}{}

\item{parentSubset}{}

\item{threshold}{(optional) Filters subsets where the average mean_gamma is greater than the threshold}

\item{antigenCol}{}

\item{stimAntigen}{}

\item{controlAntigen}{}

\item{stratifyBy}{column of pData to stratify data by}

\item{outdir}{(optional)}

\item{stratifyByValueMinuend}{(optional) stratifyBy value to be used as selector for minuend (A in A - B)}

\item{stratifyByValueSubtrahend}{(optional) stratifyBy value to be used as selector for subtrahend (B in A - B)}
}
\description{
Obtain background-corrected proportions for COMPASS subsets. Perform Wilcoxon rank sum test comparing mean for one group to mean for another group.
Defaults to only subsets above threshold 0.01.
Saves a csv with the above data to the outdir, if provided.
}
\details{
It is assumed that any given GatingSetList will have the same gates across all GatingSets within.
}
\examples{
\dontrun{
ifngResults <- compass.subset.comparisons(compassResultOrPath=myCompassResult,
       gsOrGsListOrPath=gs,
       parentSubset="4+",
       antigenCol="Antigen",
       stimAntigen="Peptide Pool 1",
       controlAntigen="DMSO",
       stratifyBy="Status",
       sampleIDCol="PATIENT ID")
}
}
