% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytokineSpecificSubsetComparison.R
\name{cytokine.specific.subset.comparison}
\alias{cytokine.specific.subset.comparison}
\title{Cytokine-specific Subset Comparison}
\usage{
cytokine.specific.subset.comparison(compassResultOrPath, gsOrGsListOrPath,
  cytokineOfInterestGate, parentSubset, threshold = 0.01, antigenCol,
  stimAntigen, controlAntigen, stratifyBy,
  showSignificanceBracket = TRUE, showTitle = TRUE, sampleIDCol,
  outdir = NULL, themeBaseSize = 15, removeGridAndBg = FALSE,
  stratifyByColors = NULL, ymax = NULL,
  stratifyByValueMinuend = NULL, stratifyByValueSubtrahend = NULL)
}
\arguments{
\item{compassResultOrPath}{}

\item{gsOrGsListOrPath}{}

\item{cytokineOfInterestGate}{}

\item{parentSubset}{}

\item{threshold}{(optional) Filters subsets where the average mean_gamma is greater than the threshold}

\item{antigenCol}{}

\item{stimAntigen}{}

\item{controlAntigen}{}

\item{stratifyBy}{column of pData to stratify plot by}

\item{showSignificanceBracket}{(optional)}

\item{showTitle}{(optional)}

\item{outdir}{(optional)}

\item{themeBaseSize}{(optional)}

\item{removeGridAndBg}{(optional)}

\item{stratifyByColors}{(optional)}

\item{ymax}{(optional)}

\item{stratifyByValueMinuend}{(optional) stratifyBy value to be used as selector for minuend (A in A - B)}

\item{stratifyByValueSubtrahend}{(optional) stratifyBy value to be used as selector for subtrahend (B in A - B)}
}
\description{
Compare cytokine-positive subsets to cytokine-negative subsets.
Additionally stratify by a column in the sample metadata.
Perform a wilcoxon rank sum test for each subset group.
}
\details{
It is assumed that any given GatingSetList will have the same gates across all GatingSets within.
}
\examples{
\dontrun{
ifngResults <- cytokine.specific.subset.comparison(compassResultOrPath=myCompassResult,
       gsOrGsListOrPath=gs,
       cytokineOfInterestGate="IFNg+",
       parentSubset="4+",
       antigenCol="Antigen",
       stimAntigen="Peptide Pool 1",
       controlAntigen="DMSO",
       stratifyBy="Status",
       showSignificanceBracket=TRUE,
       showTitle=FALSE,
       sampleIDCol="PATIENT ID")
}
}
