% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalityScoreLinePlot.R
\name{fs.line.plot}
\alias{fs.line.plot}
\title{Line plot of Functionality Score vs. some other axis}
\usage{
fs.line.plot(fsData, outdir = NULL, xaxis, yaxis, groupName, addBox = FALSE,
  ylimits = NULL, fileSuffix = NULL)
}
\arguments{
\item{fsData}{data table to plot}

\item{outdir}{(Optional) will save plot to a file in this directory, if provided}

\item{xaxis}{column in fsData for stratifying x-axis}

\item{yaxis}{column in fsData for the y-axis (probably functionality score)}

\item{groupName}{column name which identifies groups (e.g. individuals, for connecting line plots)}

\item{addBox}{(Optional) A boolean. Adds boxplots for each x-axis value if set to True.}

\item{ylimits}{(Optional) numeric vector specifying upper and lower y-axis limits}

\item{fileSuffix}{(Optional) What to add to end of filename}
}
\value{
Line plot of FS scores, unless outdir is specified
}
\description{
This function creates a line plot of Functionality Score vs. some other axis,
like those generated by the GenericCOMPASSWrapper.R script.
Some formatting of the input data table is required, so it is not very "generic".
}
\examples{
\dontrun{
# Create the fsData table from the COMPASS output FS text file.
fsFile <- "/home/path/to/CD8/8+_CMV/FS_8+_CMV.txt"
ylims <- c(0, 0.15)
fsTable <- read.table(fsFile, header=TRUE, sep="\\t", check.names=FALSE)
cellSubset <- as.character(fsTable[,"CellSubset"])
condition <- as.character(fsTable[,conditioncol])
fsData1 <- t(fsTable[,3:length(colnames(fsTable))])
fsData1 <- cbind(fsData1, ldply(rownames(fsData1), .fun = function(x) {strsplit(x, ", Time ")[[1]]}))
colnames(fsData1) <- c("FunctionalityScore", "PTID", "Time")
# Draw the line plot
fs.line.plot(fsData=fsData1, xaxis="Time", yaxis="FunctionalityScore", groupName="PTID", addBox=TRUE, ylimits=ylims)
}
}
\keyword{Functionality}
\keyword{Line}
\keyword{Plot}
