% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalityScorePlot.R
\name{fs.plot}
\alias{fs.plot}
\title{Plot of Functionality Score vs some other axis}
\usage{
fs.plot(gsOrGsListOrPath, compassResultOrPath, stratifyBy, ylims = NULL,
  outdir = NULL, plotTextExtra = "", plotWilcox = FALSE,
  themeBaseSize = 18, removeGridAndBg = FALSE, showTitle = TRUE,
  showSignificanceBracket = TRUE)
}
\arguments{
\item{gsOrGsListOrPath}{Either a GatingSet, a GatingSetList, or path to one of these objects on disk}

\item{stratifyBy}{Character vector of metadata columns to stratify plot by}

\item{outdir}{If given, plot will be saved to this directory instead of being returned in a list along with plot data}

\item{plotTextExtra}{Extra text (e.g. "CD4+, Peptide Pool 1") to place in plot title and filename. Commas and spaces, etc will be removed for filename.}
}
\description{
TODO: Make another function for paired values (i.e. line plot)
}
\examples{
\dontrun{
fsPlotList <- fs.plot(gsOrGsListOrPath="path/to/GatingSet/Folder",
       compassResultOrPath="path/to/CompassResultRDSFile.rds",
       stratifyBy="DiseaseState",
       plotTextExtra="",
       plotWilcox=TRUE)
}
}
