% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericCompassWrapper.R
\name{generic.compass.wrapper}
\alias{generic.compass.wrapper}
\title{COMPASS Wrapper}
\usage{
generic.compass.wrapper(path = NULL, seed = NULL, outdir = NULL,
  cnode = NULL, nodemarkermap = NULL, individuals = NULL,
  grouping = NULL, uniqueruns = NULL, lineplotxvar = NULL, iter = 40000,
  lineplotgroupby = NULL, countFilterThreshold = 0)
}
\arguments{
\item{path}{The path to the folder in which the GatingSetList or GatingSet is saved}

\item{seed}{(Optional) Number to set seed to [default NULL]}

\item{outdir}{Directory in which to save output, e.g. heatmaps}

\item{cnode}{Node on which to run COMPASS}

\item{nodemarkermap}{List mapping nodes to marker names}

\item{individuals}{pData column containing individual identifiers (rows of heatmap)}

\item{grouping}{(Optional) pData columns on which to group rows in heatmap, as a character vector [default NULL]}

\item{uniqueruns}{(Optional) pData column identifying unique runs. Use if you need multiple runs. [default NULL]}

\item{lineplotxvar}{(Optional) pData column which defines groups along x-axis in FS-score line plot, e.g. Time [default NULL]}

\item{iter}{(Optional) Number of COMPASS iterations to perform on each repitition (8 repetitions total) [default 40,000]}

\item{lineplotgroupby}{(Optional) This should be specified if lineplotxvar is given. pData column which defines which values to connect in the line plot (usually something like "PTID")}
}
\value{
Nothing
}
\description{
This function runs COMPASS once for each of the unique values defined by the uniqueruns argument (if provided)
}
\examples{
\dontrun{
generic.compass.wrapper(path="/home/path/to/GatingSetList",
                        seed=1,
                        outdir="/path/to/OutDirectory",
                        cnode="8+",
                        nodemarkermap=list("8+/154+" = "CD154",
                                           "8+/IFNg+" = "IFNg",
                                           "8+/IL4+" = "IL4",
                                           "8+/TNFa+" = "TNFa",
                                           "8+/IL22+" = "IL22",
                                           "8+/IL17+" = "IL17a",
                                           "8+/IL2+" = "IL2"),
                        individuals="PATIENT ID",
                        uniqueruns="Peptide")
                        }
}
\keyword{COMPASS}
