% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinePlotBooleanSubsetcellProportions.R
\name{getCatsAndSubsetNames}
\alias{getCatsAndSubsetNames}
\title{Format COMPASS Categories}
\usage{
getCatsAndSubsetNames(cats)
}
\description{
This function returns the formatted category name for each row of the categories matrix
and the ordered categories matrix
Note that it ORDERS the categories matrix columns first
Used by mergeMatricesForPlotCOMPASSResultStack function
}
