% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_create.R
\name{gs_create}
\alias{gs_create}
\title{Create a GatingSet from a directory of fcs files}
\usage{
gs_create(
  FCSDirectory,
  pattern = ".fcs$",
  recursive = FALSE,
  ncdf = TRUE,
  comp = "acquisition",
  transform = "biex",
  arcsinh_coeff = 150,
  linearChannels = 2
)
}
\arguments{
\item{FCSDirectory}{The parent directory containing the fcs files.}

\item{pattern}{a string/regexp that selects the files to read in. Defaults to
all fcs files with ".fcs$".}

\item{recursive}{Boolean specifying whether the fcs files should be searched
for within FCSDirectory recursively (i.e. are there subfolders?).}

\item{ncdf}{Boolean specifying whether the fcs files should be read in as
ncdf files.}

\item{comp}{Either "acquisition" or the filepath to a comp matrix saved as a
.csv file. Specify FALSE if you want no comp applied.}

\item{transform}{Either "biex" or "arcsinh" to apply the appropriate
transform function to all channels not defined as linear (see below).}

\item{arcsinh_coeff}{passed to the arcsinh function - defaults to 150, which
is considered appropriate for fluorescent cytometry data (use coefficient 1
if you have cytof data instead).}

\item{linearChannels}{Either a number specifying how many FSC and SSC
channels you have (don't include TIME as linear even though it is - this is
taken care of already), or a character vector naming all of the channels
you \emph{don't} want transformed (i.e. the linear channels).}
}
\value{
a GatingSet object with the comp and transformation applied as
specified.
}
\description{
This function takes a directory (or directories recursively) containing some
fcs files and reads them into the GatingSet format from flowWorkspace so that
they can be subsequentially gated on. The function can optionally read the
data in as ncdf files to save memory, can apply a compensation matrix from an
acquisition-defined slot or a provided matrix in .csv format, and can apply a
biexponential or arcsinh transform.
}
\examples{

path_to_fcs <- system.file("extdata", package = "flowHelpers")
gs <- gs_create(FCSDirectory = path_to_fcs,
pattern = ".fcs$",
recursive = FALSE,
ncdf = FALSE,
comp = FALSE,
transform = "biex",
linearChannels = )


}
