% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighlightBooleanSubsetFacsPlot.R
\name{highlight.boolean.subset.facs.plot}
\alias{highlight.boolean.subset.facs.plot}
\title{Overlay and Highlight Polyfunctional Cell Subsets on a FACS plot}
\usage{
highlight.boolean.subset.facs.plot(path, gsOrGsList = NULL, outdir = NULL,
  individualsCol, individual, conditioncol, exp, ctrl, conditioncol2 = ".",
  parentsubset, boolsubset, xaxis, yaxis, facetorder = NULL,
  facetorder2 = NULL, geomTextY = 5, geomtextX = 200)
}
\arguments{
\item{path}{path to directory holding GatingSetList or GatingSet}

\item{individualsCol}{column which defines individual}

\item{individual}{value of individual in individualsCol whose data you want to plot}

\item{conditioncol}{name of the column that defines the main experimental condition, e.g. Antigen}

\item{exp}{experimental value in conditioncol, e.g. ESAT-6}

\item{ctrl}{control value in conditioncol, e.g. DMSO}

\item{conditioncol2}{second condition on which to stratify FACS plots}

\item{parentsubset}{unique name of parent node to use for plots}

\item{boolsubset}{the full boolean subset to be used by booleanfilter()}

\item{xaxis}{a marker name to plot on the x-axis}

\item{yaxis}{a marker name to plot on the y-axis}

\item{(Optional)}{outdir saves image in output directory, if given}

\item{(Optional)}{facetorder the levels of conditioncol (e.g. Antigen) in the order you want displayed}

\item{(Optional)}{facetorder2 the levels of conditioncol2 (e.g. Time) in the order you want displayed}
}
\value{
FACS plot, unless outdir is specified
}
\description{
Defines a function to Overlay and Highlight Polyfunctional Cell Subsets on a FACS plot
This function assumes there are two columns, conditioncol and conditioncol2, upon which you are stratifying the plots
}
\examples{
\dontrun{
highlight.boolean.subset.facs.plot(path="/home/path/to/GatingSetListAllBatches",
                                   individualsCol="PTID",
                                   individual=12345678,
                                   conditioncol="Antigen",
                                   exp="ESAT-6",
                                   ctrl="DMSO",
                                   conditioncol2="Time",
                                   parentsubset="8+",
                                   boolsubset="8+/TNFa+&!8+/IFNg+&!8+/IL2+&!8+/IL4+",
                                   xaxis="TNFa",
                                   yaxis="IFNg",
                                   facetorder=c("DMSO", "ESAT-6"))
                                   }
}
\keyword{FACS}
\keyword{Plot}
\keyword{Polyfunctional}
\keyword{Subset}
