% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderHeatmapColumnsForCytokine.R
\name{orderHeatmapColumnsByCytokinePresence}
\alias{orderHeatmapColumnsByCytokinePresence}
\title{Order COMPASSresult categories columns for cytokine of interest}
\usage{
orderHeatmapColumnsByCytokinePresence(cr, cytokine,
  cats_df_cytokine_order_override = NULL, threshold = 0.01,
  minimum_dof = 1, maximum_dof = Inf)
}
\arguments{
\item{cr}{a COMPASSresult}

\item{cytokine}{The cytokine of interest}

\item{cats_df_cytokine_order_override}{Use this if you want to override the sorting of the categories matrix cytokine. Don't include "Counts".}
}
\description{
This function rearranges the mean_gamma and categories matrices so that subsets containing the cytokine of interest appear on the right.
The categories legend is rearranged such that the cytokine of interest is on the top row (so, top right cells will be shaded in).
The other cytokines are also rearranged in a way to be as aesthetically plasing as possible.
Or, specify your own cytokine order for the legend with cats_df_cytokine_order_override (doesn't override the cytokine-of-interest though).
}
\examples{
\dontrun{
new_cr <- orderHeatmapColumnsByCytokinePresence(myCompassResult, "IFNg")
# Good for the fixed_column_order = T option from modified plot function
p <- print(plot(new_cr, order_by_max_functionality = F,
                fixed_column_order = T))
grid.draw(p)
}
}
