% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareGatingSetList4Compass.R
\name{prepare.gating.set.list.4.compass}
\alias{prepare.gating.set.list.4.compass}
\title{Prepare flow data for COMPASS}
\usage{
prepare.gating.set.list.4.compass(xmlFiles = NULL, fcsFiles = NULL,
  gsDirs = NULL, gsList = list(), outDir = NULL, sampleGroups = NULL,
  keywords2import = c(), keyword4samples2exclude = NULL,
  samples2exclude = NULL)
}
\arguments{
\item{xmlFiles}{character vector (xmlFiles AND/OR gsDirs AND/OR gsList is required)}

\item{fcsFiles}{(Optional) character vector. Only needed if fcs and xml files exist in different directories.}

\item{gsDirs}{character vector (xmlFiles AND/OR gsDirs AND/OR gsList is required)}

\item{gsList}{list of GatingSet objects (xmlFiles AND/OR gsDirs AND/OR gsList is required)}

\item{outDir}{directory in which to save the final GatingSetList}

\item{sampleGroups}{(Optional) numeric vector. Specify this if the flowJo sample group is not 3 for all batches.}

\item{keywords2import}{(Optional) character vector. List of keywords to import from FlowJo workspace and into pData.}

\item{keyword4samples2exclude}{(Optional) keyword used to identify samples for exclusion}

\item{samples2exclude}{(Optional) character vector. When making GatingSetList, exclude samples whose keyword4samples2exclude column contains values in this vector. Usually a result of poor quality data, discovered during the QC step.}
}
\value{
Nothing
}
\description{
This function reads in multiple FlowJo workspace xml files and their associated FCS files.
It then combines them all into a single GatingSetList, extracting the user-provided keywords
and checking that the batches are combine-able along the way.
Optionally, you can provide a list of saved GatingSet directories (in addition to or instead of the xml files).
These are assumed to already have keywords and samples filtered.
The final GatingSetList is saved to a user-provided output directory. You can then use it directly with COMPASS or modify it further.
}
\details{
Note 1: All the batches need to have the same gating tree, so this function will drop unshared
nodes and channels from the batches to be merged. TODO: make option to turn off default
Note 2: Marker names and channel names must all be the same across the batches.
}
\examples{
\dontrun{
prepare.gating.set.list.4.compass(xmlFiles=c("/home/Batch1Files/Batch1FlowJoWorkspace.xml",
                                             "/home/Batch2Files/Batch2FlowJoWorkspace.xml",
                                             "/home/Batch3Files/Batch3FlowJoWorkspace.xml",
                                             "/home/Batch4Files/Batch4FlowJoWorkspace.xml"),
                                  outDir="/home/GatingSetListOutDir",
                                  keywords2import=c("Barcode", "Antigen", "PATIENT ID"),
                                  keyword4samples2exclude="Barcode",
                                  samples2exclude=c('1234567890', '1234567891', '1234567892', '1234567893', '1234567894'))
                                  }
}
\keyword{COMPASS}
\keyword{GatingSetList}
