% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericCompassWrapper.R
\name{run.compass.once}
\alias{run.compass.once}
\title{Run COMPASS once}
\usage{
run.compass.once(gs, cnode, individuals, nodemarkermap, iter = 40000,
  lineplotxvar = NULL, run = NULL, outdir, uniqueruns = NULL,
  grouping = NULL, lineplotgroupby = NULL, countFilterThreshold = 0,
  seed = NULL)
}
\arguments{
\item{gs}{The GatingSet of GatingSetList}

\item{cnode}{Node on which to run COMPASS}

\item{individuals}{pData column containing individual identifiers (rows of heatmap)}

\item{nodemarkermap}{List mapping nodes to marker names}

\item{iter}{(Optional) Number of COMPASS iterations to perform on each repitition (8 repetitions total) [default 40,000]}

\item{lineplotxvar}{(Optional) pData column which defines groups along x-axis in FS-score line plot, e.g. Time [default NULL]}

\item{outdir}{Directory in which to save output, e.g. heatmaps}

\item{uniqueruns}{(Optional) pData column identifying unique runs. Use if you need multiple runs. [default NULL]}

\item{grouping}{(Optional) pData columns on which to group rows in heatmap, as a character vector [default NULL]}

\item{lineplotgroupby}{(Optional) This should be specified if lineplotxvar is given. pData column which defines which values to connect in the line plot (usually something like "PTID")}

\item{seed}{(Optional) Number to set seed to [default NULL]}
}
\value{
Nothing
}
\description{
This function runs COMPASS once, saving output to disk.
Intended for use by the generic.compass.wrapper function below.
}
\keyword{COMPASS}
